+++ alicloud/extension_ess.go
@@ -2,6 +2,42 @@ package alicloud
 
 import "github.com/aliyun/alibaba-cloud-sdk-go/sdk/requests"
 
+const UserId = "userId"
+const ScalingGroup = "scaling_group"
+
+type ComparisonOperator string
+
+const (
+	Gt  = ComparisonOperator(">")
+	Gte = ComparisonOperator(">=")
+	Lt  = ComparisonOperator("<")
+	Lte = ComparisonOperator("<=")
+)
+
+type Statistics string
+
+const (
+	Avg = Statistics("Average")
+	Min = Statistics("Minimum")
+	Max = Statistics("Maximum")
+)
+
+type Period int
+
+const (
+	OneMinite     = Period(60)
+	TwoMinite     = Period(120)
+	FiveMinite    = Period(300)
+	FifteenMinite = Period(900)
+)
+
+type MetricType string
+
+const (
+	System = MetricType("system")
+	Custom = MetricType("custom")
+)
+
 type ActionResult string
 
 const (